backend/question-service/pyproject.toml
[tool.poetry]
name = "questionscraper"
version = "0.1.0"
description = ""
authors = ["blackonyyx <35429757+blackonyyx@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
SQLAlchemy = "^1.4.41"
PyMySQL = "^1.0.2"
fastapi = "^0.85.0"
uvicorn = {extras = ["standard"], version = "^0.18.3"}
gunicorn = "^20.1.0"
alembic = "^1.8.1"
psycopg2 = "^2.9.3"
beautifulsoup4 = "^4.11.1"
markdownify = "^0.11.6"
python-dotenv = "^0.21.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.1.3"
requests = "^2.28.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
